major: consolidating into one/original API
