+++ src/main/java/conseq4j/ExecutorSweepingTaskExecutionListener.java
@@ -17,7 +17,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package conseq4j.service;
+package conseq4j;
 
 import lombok.ToString;
 import lombok.extern.java.Log;
