--- src/main/java/conseq4j/service/TaskExecutionListener.java
@@ -1,40 +1,0 @@
-/*
- * The MIT License
- * Copyright 2022 Qingtian Wang.
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-
-package conseq4j.service;
-
-/**
- * Listener on each execution invocation of the target executor.
- *
- * @author Qingtian Wang
- */
-interface TaskExecutionListener {
-
-    /**
-     * @see java.util.concurrent.ThreadPoolExecutor#beforeExecute(Thread, Runnable)
-     */
-    void beforeExecute(Thread taskExecutionThread, Runnable task);
-
-    /**
-     * @see java.util.concurrent.ThreadPoolExecutor#afterExecute(Runnable, Throwable)
-     */
-    void afterExecute(Runnable task, Throwable taskExecutionError);
-
-}
