+++ src/test/java/conseq4j/ConseqTest.java
@@ -1,106 +1,217 @@
 /*
- * The MIT License
- * Copyright 2021 Qingtian Wang.
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
+ * The MIT License Copyright 2021 Qingtian Wang. Permission is hereby granted, free of charge, to
+ * any person obtaining a copy of this software and associated documentation files (the "Software"),
+ * to deal in the Software without restriction, including without limitation the rights to use,
+ * copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to the following conditions:
+ * The above copyright notice and this permission notice shall be included in all copies or
+ * substantial portions of the Software. THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY
+ * KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS
+ * BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
+ * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
+ * IN THE SOFTWARE.
  */
 package conseq4j;
 
-import org.junit.jupiter.api.Test;
+import lombok.extern.java.Log;
+import org.junit.jupiter.api.*;
 
-import java.util.UUID;
-import java.util.concurrent.Executor;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.TimeUnit;
+import java.time.Duration;
+import java.time.Instant;
+import java.util.*;
+import java.util.concurrent.*;
+import java.util.logging.ConsoleHandler;
+import java.util.logging.Handler;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import static org.junit.jupiter.api.Assertions.assertSame;
-import static org.junit.jupiter.api.Assertions.fail;
+import static java.util.stream.Collectors.toList;
+import static java.util.stream.Collectors.toSet;
+import static org.junit.jupiter.api.Assertions.*;
 
 /**
- * @author q3769
+ * @author Qingtian Wang
  */
-class ConseqTest {
+@Log public class ConseqTest {
 
-    @Test void shouldReturnSameExecutorOnSameName() {
-        UUID sequenceKey = UUID.randomUUID();
-        Conseq target = Conseq.newBuilder().build();
+    private static final int TASK_COUNT = 100;
 
-        Executor e1 = target.getSequentialExecutor(sequenceKey);
-        Executor e2 = target.getSequentialExecutor(sequenceKey);
+    private static final Level TEST_RUN_LOG_LEVEL = Level.FINE;
 
-        assertSame(e1, e2);
+    @BeforeAll public static void setLoggingLevel() {
+        Logger root = Logger.getLogger("");
+        // .level= ALL
+        root.setLevel(TEST_RUN_LOG_LEVEL);
+        for (Handler handler : root.getHandlers()) {
+            if (handler instanceof ConsoleHandler) {
+                // java.util.logging.ConsoleHandler.level = ALL
+                handler.setLevel(TEST_RUN_LOG_LEVEL);
+            }
+        }
     }
 
-    @Test void errorOnNonPositiveConcurrency() {
-        int errors = 0;
-        try {
-            Conseq.newBuilder().globalConcurrency(0);
-        } catch (IllegalArgumentException e) {
-            errors++;
-        }
-        try {
-            Conseq.newBuilder().globalConcurrency(-999);
-        } catch (IllegalArgumentException e) {
-            errors++;
-        }
-        if (errors == 2)
-            return;
-        fail();
-    }
-
-    @Test void errorOnNonPositiveTaskQueueSize() {
-        int errors = 0;
-        try {
-            Conseq.newBuilder().executorTaskQueueSize(0);
-        } catch (IllegalArgumentException e) {
-            errors++;
-        }
-        try {
-            Conseq.newBuilder().executorTaskQueueSize(-999);
-        } catch (IllegalArgumentException e) {
-            errors++;
-        }
-        if (errors == 2)
-            return;
-        fail();
-    }
-
-    @Test void irrevocable() {
-        Conseq target = Conseq.newBuilder().build();
-        final ExecutorService sequentialExecutor = target.getSequentialExecutor("foo");
-        sequentialExecutor.execute(() -> {
+    private static void awaitAllComplete(List<Future<SpyingTask>> futures) {
+        futures.forEach(f -> {
             try {
-                TimeUnit.SECONDS.sleep(1L);
-            } catch (InterruptedException ex) {
-                Logger.getLogger(ConseqTest.class.getName()).log(Level.SEVERE, null, ex);
-                Thread.currentThread().interrupt();
+                f.get();
+            } catch (InterruptedException | ExecutionException ex) {
+                throw new IllegalStateException(ex);
             }
         });
+    }
+
+    private static List<SpyingTask> createSpyingTasks(int total) {
+        List<SpyingTask> result = new ArrayList<>();
+        for (int i = 0; i < total; i++) {
+            result.add(new SpyingTask(i));
+        }
+        return result;
+    }
+
+    @BeforeEach void setUp(TestInfo testInfo) {
+        log.info(String.format("================================== start test: %s", testInfo.getDisplayName()));
+    }
+
+    @AfterEach void tearDown(TestInfo testInfo) {
+        log.info(String.format("================================== done test: %s", testInfo.getDisplayName()));
+    }
+
+    @Test void concurrencyBoundedByTotalTaskCount() {
+        Conseq defaultConseq = Conseq.newBuilder().build();
+
+        List<Future<SpyingTask>> futures = new ArrayList<>();
+        createSpyingTasks(TASK_COUNT).forEach(task -> futures.add(
+                defaultConseq.getSequentialExecutor(UUID.randomUUID()).submit((Callable<SpyingTask>) task)));
+
+        final long totalRunThreads = toDoneTasks(futures).stream().map(SpyingTask::getRunThreadName).distinct().count();
+        log.log(Level.INFO, "{0} tasks were run by {1} threads", new Object[] { TASK_COUNT, totalRunThreads });
+        assertTrue(totalRunThreads <= TASK_COUNT);
+    }
+
+    @Test void concurrencyBoundedByMaxConcurrency() {
+        List<SpyingTask> sameTasks = createSpyingTasks(TASK_COUNT);
+        Conseq lowConcurrencyService = Conseq.newBuilder().globalConcurrency(TASK_COUNT / 10).build();
+        List<Future<SpyingTask>> lowConcurrencyFutures = new ArrayList<>();
+        long lowConcurrencyStart = System.nanoTime();
+        sameTasks.forEach(task -> lowConcurrencyFutures.add(
+                lowConcurrencyService.getSequentialExecutor(UUID.randomUUID()).submit((Callable<SpyingTask>) task)));
+        awaitAllComplete(lowConcurrencyFutures);
+        long lowConcurrencyTime = System.nanoTime() - lowConcurrencyStart;
+
+        Conseq highConcurrencyService = Conseq.newBuilder().globalConcurrency(TASK_COUNT).build();
+        List<Future<SpyingTask>> highConcurrencyFutures = new ArrayList<>();
+        long highConcurrencyStart = System.nanoTime();
+        sameTasks.forEach(task -> highConcurrencyFutures.add(
+                highConcurrencyService.getSequentialExecutor(UUID.randomUUID()).submit((Callable<SpyingTask>) task)));
+        awaitAllComplete(highConcurrencyFutures);
+        long highConcurrencyTime = System.nanoTime() - highConcurrencyStart;
+
+        log.log(Level.INFO, "Low concurrency run time {0}, high concurrency run time {1}",
+                new Object[] { Duration.ofNanos(lowConcurrencyTime), Duration.ofNanos(highConcurrencyTime) });
+        assertHighConcurrencyIsFaster(lowConcurrencyTime, highConcurrencyTime);
+    }
+
+    void assertHighConcurrencyIsFaster(long lowConcurrencyTime, long highConcurrencyTime) {
+        assertTrue(lowConcurrencyTime > highConcurrencyTime);
+    }
+
+    @Test void invokeAllRunsTasksOfSameSequenceKeyInSequence() throws InterruptedException {
+        ConcurrentSequencer defaultConseq = Conseq.newBuilder().build();
+        List<SpyingTask> tasks = createSpyingTasks(TASK_COUNT);
+        UUID sameSequenceKey = UUID.randomUUID();
+
+        log.log(Level.INFO, () -> "Start single sync invoke all " + tasks.size() + " tasks under same sequence key "
+                + sameSequenceKey);
+        final List<Future<SpyingTask>> futures = defaultConseq.getSequentialExecutor(sameSequenceKey).invokeAll(tasks);
+        log.log(Level.INFO, () -> "Done single sync invoke all " + tasks.size() + " tasks under same sequence key "
+                + sameSequenceKey);
+
+        final List<SpyingTask> doneTasks = toDoneTasks(futures);
+        assertSingleThread(doneTasks);
+        assertSequence(doneTasks);
+    }
+
+    @Test void invokeAnyChoosesTaskInSequenceRange() throws InterruptedException, ExecutionException {
+        ConcurrentSequencer defaultConseq = Conseq.newBuilder().build();
+        List<SpyingTask> tasks = createSpyingTasks(TASK_COUNT);
+        UUID sameSequenceKey = UUID.randomUUID();
+
+        log.log(Level.INFO, () -> "Start single sync invoke any in " + tasks.size() + " tasks under same sequence key "
+                + sameSequenceKey);
+        SpyingTask doneTask = defaultConseq.getSequentialExecutor(sameSequenceKey).invokeAny(tasks);
+        log.log(Level.INFO, () -> "Done single sync invoke any in " + tasks.size() + " tasks under same sequence key "
+                + sameSequenceKey);
+
+        final Integer scheduledSequence = doneTask.getScheduledSequence();
+        log.log(Level.INFO, "Chosen task sequence : {0}", scheduledSequence);
+        assertTrue(scheduledSequence >= 0 && scheduledSequence < TASK_COUNT);
+    }
+
+    @Test void submitsRunAllTasksOfSameSequenceKeyInSequence() throws InterruptedException {
+        ConcurrentSequencer defaultConseq = Conseq.newBuilder().build();
+        List<SpyingTask> tasks = createSpyingTasks(TASK_COUNT);
+        UUID sameSequenceKey = UUID.randomUUID();
+        final int extraFactorEnsuringAllDone = TASK_COUNT / 10;
+        final int timeToAllowAllComplete = (TASK_COUNT + extraFactorEnsuringAllDone) * SpyingTask.MAX_RUN_TIME_MILLIS;
+
+        log.log(Level.INFO, () -> "Start async submitting each of " + tasks.size() + " tasks under same sequence key "
+                + sameSequenceKey);
+        tasks.forEach(task -> defaultConseq.getSequentialExecutor(sameSequenceKey).execute(task));
+        log.log(Level.INFO, () -> "Done async submitting each of " + tasks.size() + " tasks under same sequence key "
+                + sameSequenceKey);
+        TimeUnit.MILLISECONDS.sleep(timeToAllowAllComplete);
+
+        assertSingleThread(tasks);
+        assertSequence(tasks);
+    }
+
+    @Test void excessiveTasksOverTaskQueueCapacityWillBeRejected() {
+        int executorTaskQueueCapacity = 42;
+        assertTrue(executorTaskQueueCapacity < TASK_COUNT);
+        ConcurrentSequencer taskQueueCapacityLimited =
+                Conseq.newBuilder().executorTaskQueueCapacity(executorTaskQueueCapacity).build();
+        List<SpyingTask> tasks = createSpyingTasks(TASK_COUNT);
+        UUID sameSequenceKey = UUID.randomUUID();
 
         try {
-            sequentialExecutor.shutdown();
-        } catch (UnsupportedOperationException ex) {
-            try {
-                sequentialExecutor.shutdownNow();
-            } catch (UnsupportedOperationException ex2) {
-                return;
-            }
+            tasks.forEach(task -> taskQueueCapacityLimited.getSequentialExecutor(sameSequenceKey).execute(task));
+        } catch (RejectedExecutionException e) {
+            log.log(Level.WARNING,
+                    "almost never a good idea to limit task queue capacity; unless you intend to reject excessive tasks that the task queue cannot hold, consider default/unbounded capacity instead",
+                    e);
+            return;
         }
         fail();
+    }
+
+    void assertSingleThread(List<SpyingTask> tasks) {
+        Set<String> uniqueThreadNames =
+                tasks.stream().map(SpyingTask::getRunThreadName).filter(Objects::nonNull).collect(toSet());
+        assertEquals(1, uniqueThreadNames.size());
+        log.log(Level.INFO, "{0} tasks executed by single thread {1}",
+                new Object[] { tasks.size(), uniqueThreadNames.stream().findFirst().get() });
+    }
+
+    List<SpyingTask> toDoneTasks(List<Future<SpyingTask>> futures) {
+        log.log(Level.FINER, () -> "Wait and get all results on futures " + futures);
+        final List<SpyingTask> doneTasks = futures.stream().map(f -> {
+            try {
+                return f.get();
+            } catch (InterruptedException | ExecutionException ex) {
+                throw new IllegalStateException(ex);
+            }
+        }).collect(toList());
+        log.log(Level.FINER, () -> "All futures done, results: " + doneTasks);
+        return doneTasks;
+    }
+
+    void assertSequence(List<SpyingTask> tasks) {
+        for (int i = 0; i < tasks.size() - 1; i++) {
+            final Instant currentEnd = tasks.get(i).getRunEnd();
+            final Instant nextStart = tasks.get(i + 1).getRunStart();
+            assertTrue(currentEnd.isBefore(nextStart));
+        }
+        log.log(Level.INFO, "{0} tasks executed sequentially in chronological order", tasks.size());
     }
 }
