+++ src/test/java/conseq4j/SpyingTask.java
@@ -12,7 +12,7 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
  * IN THE SOFTWARE.
  */
-package conseq4j.service;
+package conseq4j;
 
 import lombok.Getter;
 import lombok.ToString;
